Extract review test-case loader
