+++ scinoephile/cli/audit/audit_review_cli_base.py
@@ -5,8 +5,6 @@
 from __future__ import annotations
 
 from argparse import ArgumentParser
-from collections.abc import Sequence
-from pathlib import Path
 from typing import ClassVar
 
 from scinoephile.analysis.audit.review import ReviewAuditFilter
@@ -14,16 +12,10 @@ from scinoephile.common.argument_parsing import (
     enum_arg,
     get_arg_groups_by_name,
 )
-from scinoephile.core.llms import TestCase
-from scinoephile.llms.review import ReviewManager
 
 from .audit_cli_base import AuditCliBase
-from .utils import load_audit_test_cases
 
-__all__ = [
-    "AuditReviewCliBase",
-    "load_review_test_cases",
-]
+__all__ = ["AuditReviewCliBase"]
 
 AUDIT_REVIEW_CLI_LOCALIZATIONS: dict[str, dict[str, str]] = {
     "zh-hans": {
@@ -107,25 +99,3 @@ class AuditReviewCliBase(AuditCliBase):
                 "traditional variants are included automatically"
             ),
         )
-
-
-def load_review_test_cases(
-    parser: ArgumentParser,
-    json_path: Path | None,
-) -> Sequence[TestCase]:
-    """Load optional review test cases from JSON.
-
-    Arguments:
-        parser: parser used to report user-facing errors
-        json_path: optional review JSON path
-    Returns:
-        loaded review test cases
-    """
-    if json_path is None:
-        return ()
-    return load_audit_test_cases(
-        parser,
-        json_path,
-        ReviewManager,
-        workflow_name="review",
-    )
