+++ scinoephile/cli/audit/audit_review_dual_cli.py
@@ -17,8 +17,12 @@ from scinoephile.common.argument_parsing import (
 from scinoephile.core import ScinoephileError
 from scinoephile.lang.zho.script.conversion import get_zho_character_variants
 
-from .audit_review_cli_base import AuditReviewCliBase, load_review_test_cases
-from .utils import validate_subtitle_index_range, write_audit_report
+from .audit_review_cli_base import AuditReviewCliBase
+from .utils import (
+    load_review_test_cases,
+    validate_subtitle_index_range,
+    write_audit_report,
+)
 
 __all__ = ["AuditReviewDualCli"]
