+++ scinoephile/cli/audit/audit_review_trad_cli.py
@@ -18,8 +18,12 @@ from scinoephile.common.argument_parsing import get_arg_groups_by_name, input_fi
 from scinoephile.core import ScinoephileError
 from scinoephile.lang.zho.script.conversion import get_zho_character_variants
 
-from .audit_review_cli_base import AuditReviewCliBase, load_review_test_cases
-from .utils import validate_subtitle_index_range, write_audit_report
+from .audit_review_cli_base import AuditReviewCliBase
+from .utils import (
+    load_review_test_cases,
+    validate_subtitle_index_range,
+    write_audit_report,
+)
 
 __all__ = ["AuditReviewTradCli"]
